refactor(dfdaemon): rename package daemon to dfdaemon
