+++ dfdaemon/global/global.go
