+++ dfdaemon/initializer/initializer.go
@@ -29,10 +29,10 @@ import (
 
 	log "github.com/Sirupsen/logrus"
 
-	"github.com/alibaba/Dragonfly/daemon/constant"
-	. "github.com/alibaba/Dragonfly/daemon/global"
-	. "github.com/alibaba/Dragonfly/daemon/muxconf"
-	"github.com/alibaba/Dragonfly/daemon/util"
+	"github.com/alibaba/Dragonfly/dfdaemon/constant"
+	. "github.com/alibaba/Dragonfly/dfdaemon/global"
+	. "github.com/alibaba/Dragonfly/dfdaemon/muxconf"
+	"github.com/alibaba/Dragonfly/dfdaemon/util"
 )
 
 func init() {
